Re-insert afterpay.js script tag in header
Previously, we were injecting the afterpay.js script tag from the mobile app using JavaScript. Before that, we had put it in a script tag in the body.

The best way to do it will be to use a script tag in the header. We want it in the header so that it will load straight away.