Move widget bootstrap script to its own file
Previously, we loaded the HTML with the widget bootstrap in it. Now we want to have the HTML embedded in the app, so that the app can change it depending on the environment. We still want to host the bootstrap centrally, however.

We’re leaving the HTML bootstrap around for now for backwards compatibility. We’ll remove it very soon.