Forward checkout errors to client
Sometimes, errors in the checkout occur. When they do, the response was from our example API was a simple `{}` which was not very informative.

Now we send back the error code and message from the checkout API along to the client. What the client wants to do it with it is up to them.